Game コンポーネントを分離
